CharacterBoardView 모놀리스 분해 — 1,856줄 → 296줄 (CQ-1, 순수 이동)
- characters/ 아래 7개 컴포넌트 파일 분리(StageRail·TagChips·CharacterCard·FeaturedImageSlot·CostumeDetail·CharacterDetailModal·AddCharacterModal) + claimReactKey 유틸
- 동작 변경 0 — 원본 24개 블록 바이트 단위 일치 검증(export 정규화 제외), 뷰 추가 라인은 import 6줄뿐
- 테스트 앵커 ~70개를 실제 파일로 재배치(삭제 0), 전역 부재 검사는 결합 소스(characterBoardAll)로 유지·강화
- 게이트: typecheck / node --test 46 pass / build:vite 통과
