+++ src/utils/claimReactKey.ts
@@ -1,0 +1,7 @@
+import type { KeyboardEvent as ReactKeyboardEvent } from 'react';
+
+export function claimReactKey(event: ReactKeyboardEvent<HTMLElement>) {
+  event.preventDefault();
+  event.stopPropagation();
+  event.nativeEvent.stopImmediatePropagation?.();
+}
